--- src/ServerApi/Services/Adnc.Usr/Adnc.Usr.WebApi/Authorize/Tokens.cs
@@ -1,8 +1,0 @@
-﻿namespace Adnc.Usr.WebApi.Authorize
-{
-    public enum Tokens
-    {
-        AccessToken = 1,
-        RefreshToken = 2
-    }
-}
