+++ src/ServerApi/Services/Adnc.Usr/Adnc.Usr.WebApi/GlobalUsings.cs
@@ -1,10 +1,11 @@
 ﻿global using Adnc.Infra.Core.Configuration;
+global using Adnc.Infra.Helper;
 global using Adnc.Shared;
 global using Adnc.Shared.Application.Contracts.Dtos;
 global using Adnc.Shared.Consts.Permissions.Usr;
+global using Adnc.Shared.WebApi.Authentication.JwtBearer;
 global using Adnc.Usr.Application.Contracts.Dtos;
 global using Adnc.Usr.Application.Contracts.Services;
-global using Adnc.Usr.WebApi.Authorize;
 global using Microsoft.AspNetCore.Authorization;
 global using Microsoft.AspNetCore.Builder;
 global using Microsoft.AspNetCore.Hosting;
@@ -13,8 +14,3 @@ global using Microsoft.AspNetCore.Mvc;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Options;
-global using Microsoft.IdentityModel.Tokens;
-global using System.IdentityModel.Tokens.Jwt;
-global using System.Security.Claims;
-global using System.Text;
-
