+++ src/ServerApi/Services/Shared/Adnc.Shared.WebApi/Authentication/Basic/BasicTokenHelper.cs
@@ -1,6 +1,6 @@
 ﻿namespace Adnc.Shared.WebApi.Authentication.Basic;
 
-public static class BasicSecurityTokenHandlerExtension
+public static class BasicTokenHelper
 {
     public static Func<BasicTokenValidatedContext, Task> TokenValidatedDelegate =>
         (context) =>
