+++ src/ServerApi/Services/Shared/Adnc.Shared.WebApi/Authentication/JwtBearer/Tokens.cs
@@ -1,0 +1,7 @@
+﻿namespace Adnc.Shared.WebApi.Authentication.JwtBearer;
+
+public enum Tokens
+{
+    AccessToken = 1,
+    RefreshToken = 2
+}
